branche Domoteek-server -- ajout des class des tables de la bdd + debut traitement message I_HAVE_NO_SERVER_ID

+++ Domoteek_server/src/tables/Boite_envoi.java
@@ -2,25 +2,25 @@ package tables;
 
 public class Boite_envoi {
 	
-	private int boite_envoi_id ;
-    private int objet_id ;
+	private byte boite_envoi_id ;
+    private byte objet_id ;
     
-    public Boite_envoi(int boite_envoi_id, int objet_id) {
+    public Boite_envoi(byte boite_envoi_id, byte objet_id) {
     	this.boite_envoi_id = boite_envoi_id;
     	this.objet_id = objet_id ;
     }
     
-    public int getBoite_envoi_id() {
+    public byte getBoite_envoi_id() {
 		return boite_envoi_id ;
 	}
-	public void setBoite_envoi_id(int boite_envoi_id ) {
+	public void setBoite_envoi_id(byte boite_envoi_id ) {
 		this.boite_envoi_id = boite_envoi_id ;
 	}
     
-    public int getObjet_id() {
+    public byte getObjet_id() {
 		return objet_id ;
 	}
-	public void setObjet_id(int objet_id ) {
+	public void setObjet_id(byte objet_id ) {
 		this.objet_id = objet_id ;
 	}
 }
